Split hardware interface to separate file
